refactor: move bounce model into module
